+++ pandas_datareader/commons.py
@@ -1,10 +1,17 @@
+import time
 import warnings
 import numpy as np
+import datetime as dt
+
+from pandas import to_datetime
 import pandas.compat as compat
 from pandas.core.common import PandasError
 from pandas import Panel, DataFrame
+from pandas.io.common import urlopen
+from pandas import read_csv
+from pandas.compat import StringIO, bytes_to_str
+from pandas.util.testing import _network_error_classes
 
-from pandas_datareader.commons.date_chunks import _in_chunks
 
 class SymbolWarning(UserWarning):
     pass
@@ -56,3 +63,57 @@ def _dl_mult_symbols(symbols, start, end, interval, chunksize, retry_count, paus
         # cannot construct a panel with just 1D nans indicating no data
         raise RemoteDataError("No data fetched using "
                               "{0!r}".format(method.__name__))
+
+
+def _sanitize_dates(start, end):
+    """
+    Return (datetime_start, datetime_end) tuple
+    if start is None - default is 2010/01/01
+    if end is None - default is today
+    """
+    start = to_datetime(start)
+    end = to_datetime(end)
+    if start is None:
+        start = dt.datetime(2010, 1, 1)
+    if end is None:
+        end = dt.datetime.today()
+    return start, end
+
+def _in_chunks(seq, size):
+    """
+    Return sequence in 'chunks' of size defined by size
+    """
+    return (seq[pos:pos + size] for pos in range(0, len(seq), size))
+
+def _retry_read_url(url, retry_count, pause, name):
+    """
+    Open url (and retry)
+    """
+    for _ in range(retry_count):
+        time.sleep(pause)
+
+        # kludge to close the socket ASAP
+        try:
+            with urlopen(url) as resp:
+                lines = resp.read()
+        except _network_error_classes:
+            pass
+        else:
+            rs = read_csv(StringIO(bytes_to_str(lines)), index_col=0,
+                          parse_dates=True, na_values='-')[::-1]
+            # Yahoo! Finance sometimes does this awesome thing where they
+            # return 2 rows for the most recent business day
+            if len(rs) > 2 and rs.index[-1] == rs.index[-2]:  # pragma: no cover
+                rs = rs[:-1]
+
+            #Get rid of unicode characters in index name.
+            try:
+                rs.index.name = rs.index.name.decode('unicode_escape').encode('ascii', 'ignore')
+            except AttributeError:
+                #Python 3 string has no decode method.
+                rs.index.name = rs.index.name.encode('ascii', 'ignore').decode()
+
+            return rs
+
+    raise IOError("after %d tries, %s did not "
+                  "return a 200 for url %r" % (retry_count, name, url))
