--- pandas_datareader/commons/date_chunks.py
@@ -1,22 +1,0 @@
-import datetime as dt
-from pandas import to_datetime
-
-def _sanitize_dates(start, end):
-    """
-    Return (datetime_start, datetime_end) tuple
-    if start is None - default is 2010/01/01
-    if end is None - default is today
-    """
-    start = to_datetime(start)
-    end = to_datetime(end)
-    if start is None:
-        start = dt.datetime(2010, 1, 1)
-    if end is None:
-        end = dt.datetime.today()
-    return start, end
-
-def _in_chunks(seq, size):
-    """
-    Return sequence in 'chunks' of size defined by size
-    """
-    return (seq[pos:pos + size] for pos in range(0, len(seq), size))
